Refactor justfile and add .editorconfig file

+++ justfile
@@ -1,6 +1,7 @@
+[private]
 ln-bib:
-  mkdir -p files/bibliography
-  ln -s ~/projects/share/bibliography/research.bib files/bibliography/research.bib
+    mkdir -p files/bibliography
+    [ -e files/bibliography/research.bib ] || ln -s ~/projects/share/bibliography/research.bib files/bibliography/research.bib
 
 sync:
-  rsync --update --recursive ~/projects/share/quarto/ ./
+    rsync --update --recursive ~/projects/share/quarto/ ./
